+++ src/mo_def/syntax.ml
@@ -163,6 +163,22 @@ type sugar = bool (* Is the source of a function body a block `<block>`,
                      This flag is used to correctly desugar an actor's
                      public functions as oneway, shared functions *)
 
+type loop_flags = { mutable has_break : bool; mutable has_continue : bool }
+
+let new_loop_flags () : loop_flags = { has_break = false; has_continue = false }
+
+type control = Break | Continue
+
+let auto_s = "<>auto"
+let auto_continue_s = "continue <>auto"
+
+let break_label kind (id_opt : id option) =
+  match kind, id_opt with
+  | Break, None -> auto_s
+  | Continue, None -> auto_continue_s
+  | _, Some {Source.it; _} -> it
+
+
 type id_ref = (string, mut' * exp option) Source.annotated_phrase
 and hole_sort = Named of string | Anon of int
 and exp = (exp', typ_note) Source.annotated_phrase
@@ -198,11 +214,11 @@ and exp' =
   | OrE of exp * exp                           (* disjunction *)
   | IfE of exp * exp * exp                     (* conditional *)
   | SwitchE of exp * case list                 (* switch *)
-  | WhileE of exp * exp                        (* while-do loop *)
-  | LoopE of exp * exp option                  (* do-while loop *)
-  | ForE of pat * exp * exp                    (* iteration *)
+  | WhileE of exp * exp * loop_flags       (* while-do loop *)
+  | LoopE of exp * exp option * loop_flags (* do-while loop *)
+  | ForE of pat * exp * exp * loop_flags   (* iteration *)
   | LabelE of id * typ * exp                   (* label *)
-  | BreakE of id * exp                         (* break *)
+  | BreakE of control * id option * exp        (* break *)
   | RetE of exp                                (* return *)
   | DebugE of exp                              (* debugging *)
   | AsyncE of exp option * async_sort * typ_bind * exp (* future / computation *)